Add upgrade logic to handle MicroK8sClusterUpgradeEvent to microk8scluster.py
This commit adds an upgrade method to the MicroK8sClusterEvent class that handles the MicroK8sClusterUpgradeEvent. When an instance of this event is received, the handle_event method checks if it is an instance of MicroK8sClusterUpgradeEvent and calls the upgrade method, passing in the upgrade version as an argument. The upgrade method can be used to perform the necessary upgrade logic specific to the upgrade version provided. At present, the method does not perform any upgrade logic, but the developer can add the required logic here in the future.